# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.CreateInput:
    properties:
      name:
        type: string
      sort:
        type: integer
      values:
        items:
          $ref: '#/definitions/handler.SpecValueInput'
        type: array
    required:
    - name
    type: object
...
  handler.SkU:
    properties:
      price:
        description: 价格
        type: integer
      specItems:
        description: SKUCode 不收客户端值，由服务端从规格组合派生，创建后在响应中返回
        items:
          $ref: '#/definitions/handler.specItemReq'
        minItems: 1
        type: array
      stock:
        description: 库存
        minimum: 0
        type: integer
    required:
    - price
    - specItems
    type: object
  handler.SpecValueInput:
    properties:
      name:
        type: string
      sort:
        type: integer
    required:
    - name
    type: object
  handler.UpdateInput:
    properties:
      name:
        type: string
      sort:
        type: integer
      values:
        items:
          $ref: '#/definitions/handler.UpdateValueInput'
        type: array
    required:
    - name
    type: object
...
  handler.UpdateValueInput:
    properties:
      name:
        type: string
      sort:
        type: integer
      specValueId:
        description: 规格值ID，0 表示新增
        type: integer
    required:
    - name
    type: object
...
  handler.createProductReq:
    properties:
      categoryId:
        description: 分类ID
        type: integer
      desc:
        description: 商品描述
        type: string
      name:
        description: 商品名称
        type: string
      skus:
        description: 商品规格
        items:
          $ref: '#/definitions/handler.SkU'
        type: array
    required:
    - categoryId
    - name
    - skus
    type: object
...
  handler.specItemReq:
    properties:
      specId:
        description: 规格维度ID
        type: integer
      valueId:
        description: 规格值ID
        type: integer
    required:
    - specId
    - valueId
    type: object
  handler.updateProductReq:
    properties:
      categoryId:
        type: integer
      desc:
        type: string
      name:
        type: string
      productId:
        type: integer
      urls:
        items:
          type: string
        type: array
    required:
    - productId
    type: object
...
  productapp.ProductDTO:
    properties:
      categoryId:
        type: integer
      createdAt:
        type: string
      desc:
        type: string
      name:
        type: string
      productId:
        type: integer
      skus:
        items:
          $ref: '#/definitions/productapp.SKUDTO'
        type: array
      status:
        type: integer
      updatedAt:
        type: string
    type: object
  productapp.ProductListResult:
    properties:
      list:
        items:
          $ref: '#/definitions/productapp.ProductDTO'
        type: array
      total:
        type: integer
    type: object
  productapp.SKUDTO:
    properties:
      price:
        type: integer
      skuCode:
        type: string
      specDesc:
        description: 展示文案，如 "红色 / L码"
        type: string
      specItems:
        items:
          $ref: '#/definitions/productapp.SpecItemDTO'
        type: array
      stock:
        type: integer
    type: object
  productapp.SpecItemDTO:
    properties:
      specId:
        type: integer
      specName:
        type: string
      valueId:
        type: integer
      valueName:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  specapp.ListDto:
    properties:
      list:
        items:
          $ref: '#/definitions/specapp.SpecDTO'
        type: array
      total:
        type: integer
    type: object
  specapp.SpecDTO:
    properties:
      createdAt:
        type: string
      name:
        description: 规格名称
        type: string
      sort:
        type: integer
      specId:
        type: integer
      updatedAt:
        type: string
      values:
        items:
          $ref: '#/definitions/specapp.SpecValueDTO'
        type: array
    type: object
  specapp.SpecValueDTO:
    properties:
      name:
        type: string
      sort:
        type: integer
      specId:
        type: integer
      specValueId:
        type: integer
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: DDD 架构示例服务（用户 / 商品目录 / 订单）。host 为网关端口，经网关按前缀转发到各服务
  title: Demo Service API
  version: "1.0"
paths:
...
  /product/create:
    post:
      consumes:
      - application/json
      description: 创建商品及其 SKU；SKU 编码和规格名称由服务端生成
      parameters:
      - description: 商品信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/productapp.ProductDTO'
              type: object
        "400":
          description: 参数校验失败或商品信息无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建商品
      tags:
      - 商品
  /product/list:
    post:
      description: 根据分类、商品ID或名称筛选商品并分页返回
      parameters:
      - description: 分类ID
        in: query
        name: categoryId
        type: integer
      - description: 商品ID
        in: query
        name: productId
        type: integer
      - description: 商品名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/productapp.ProductListResult'
              type: object
        "400":
          description: 查询参数校验失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 分页查询商品
      tags:
      - 商品
  /product/update:
    post:
      consumes:
      - application/json
      description: 根据商品ID更新商品的分类、名称、描述和图片地址
      parameters:
      - description: 商品更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/productapp.ProductDTO'
              type: object
        "400":
          description: 参数校验失败或商品信息无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新商品
      tags:
      - 商品
  /spec/delete/{id}:
    delete:
      description: 根据规格ID删除规格维度
      parameters:
      - description: 规格ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 规格ID格式错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除规格
      tags:
      - 规格
  /spec/findById/{id}:
    get:
      description: 根据规格ID查询规格详情（含规格值列表）
      parameters:
      - description: 规格ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/specapp.SpecDTO'
              type: object
        "400":
          description: 规格ID格式错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 根据ID查询规格
      tags:
      - 规格
  /spec/findByName/{name}:
    get:
      description: 根据规格名称精确查询规格详情（含规格值列表）
      parameters:
      - description: 规格名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/specapp.SpecDTO'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 根据名称查询规格
      tags:
      - 规格
  /spec/list:
    get:
      description: 分页查询规格维度列表，支持按名称模糊搜索；返回列表与总数
      parameters:
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10，最大100）
        in: query
        name: pageSize
        type: integer
      - description: 规格名称（模糊搜索）
        in: query
        name: name
        type: string
      - description: 规格ID
        in: query
        name: specId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/specapp.ListDto'
              type: object
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 分页查询规格列表
      tags:
      - 规格
  /spec/register:
    post:
      consumes:
      - application/json
      description: 创建一个新的规格维度（如颜色、尺寸等），可同时传入规格值列表一次性创建；同名规格已存在时返回错误
      parameters:
      - description: 规格信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/specapp.SpecDTO'
              type: object
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建规格
      tags:
      - 规格
  /spec/update/{id}:
    put:
      consumes:
      - application/json
      description: 根据规格ID更新规格维度及规格值。Values 须传该规格的完整值列表：带ID的值被更新、不带ID的值被新增、库中缺席的值被删除；单事务保证一致
      parameters:
      - description: 规格ID
        in: path
        name: id
        required: true
        type: integer
      - description: 规格信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/specapp.SpecDTO'
              type: object
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 整包更新规格
      tags:
      - 规格
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"